producer/requirements.txt
redis
markovify
kafka-python
